component-4-quality/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
asyncpg>=0.29.0

# Neo4j Graph Database
neo4j==5.15.0

# Machine Learning
scikit-learn>=1.3.2
xgboost>=2.0.3
pandas>=2.1.3
numpy>=1.26.2
joblib>=1.3.2

# NLP & Semantic Analysis
sentence-transformers>=2.2.2
spacy>=3.7.2
transformers>=4.35.2
torch>=2.1.1

# Report Generation
openpyxl==3.1.2
reportlab==4.0.7
xlsxwriter==3.1.9

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.1
httpx>=0.25.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
